refactor: Move customOp and type definitions to their own modules
