+++ tket2-py/tket2/_tket2/ops.pyi
@@ -1,6 +1,8 @@
 from enum import Enum
 from typing import Any, Iterable
 
+from tket2._tket2.types import HugrType
+
 class Tk2Op(Enum):
     """A rust-backed Tket2 built-in operation."""
 
@@ -10,6 +12,9 @@ class Tk2Op(Enum):
     @staticmethod
     def values() -> Iterable[Tk2Op]:
         """Iterate over all operation variants."""
+
+    def to_custom(self) -> CustomOp:
+        """Convert to a custom operation."""
 
     @property
     def name(self) -> str:
@@ -58,3 +63,22 @@ class Pauli(Enum):
         """Get the string name of the Pauli."""
 
     def __eq__(self, value: Any) -> bool: ...
+
+class CustomOp:
+    """A HUGR custom operation."""
+
+    def __init__(
+        self,
+        extension: str,
+        op_name: str,
+        input_types: list[HugrType],
+        output_types: list[HugrType],
+    ) -> None:
+        """Create a new custom operation from name and input/output types."""
+
+    def to_custom(self) -> CustomOp:
+        """Convert to a custom operation. Identity operation."""
+
+    @property
+    def name(self) -> str:
+        """Fully qualified (including extension) name of the operation."""
