+++ tket2-py/tket2/circuit/__init__.py
@@ -6,8 +6,6 @@ from .._tket2.circuit import (
     Node,
     Wire,
     CircuitCost,
-    CustomOp,
-    HugrType,
     validate_circuit,
     render_circuit_dot,
     render_circuit_mermaid,
@@ -30,8 +28,6 @@ __all__ = [
     "Node",
     "Wire",
     "CircuitCost",
-    "CustomOp",
-    "HugrType",
     "validate_circuit",
     "render_circuit_dot",
     "render_circuit_mermaid",
